+++ cgateweb/src/haDiscovery.js
@@ -38,7 +38,8 @@ const {
     HA_ORIGIN_SW_VERSION,
     HA_ORIGIN_SUPPORT_URL,
     CGATE_CMD_TREEXML,
-    NEWLINE
+    NEWLINE,
+    entityIdFields
 } = require('./constants');
 
 class HaDiscovery {
@@ -406,7 +407,7 @@ class HaDiscovery {
             const payload = {
                 name: null,
                 unique_id: uniqueId,
-                ...(entityId && { default_entity_id: `${HA_COMPONENT_LIGHT}.${entityId}`, object_id: entityId }),
+                ...(entityId && entityIdFields(HA_COMPONENT_LIGHT, entityId)),
                 state_topic: `${MQTT_TOPIC_PREFIX_READ}/${networkId}/${appId}/${groupId}/${MQTT_TOPIC_SUFFIX_STATE}`,
                 command_topic: `${MQTT_TOPIC_PREFIX_WRITE}/${networkId}/${appId}/${groupId}/${MQTT_CMD_TYPE_RAMP}`,
                 brightness_state_topic: `${MQTT_TOPIC_PREFIX_READ}/${networkId}/${appId}/${groupId}/${MQTT_TOPIC_SUFFIX_LEVEL}`,
@@ -518,7 +519,7 @@ class HaDiscovery {
         const payload = {
             name: null,
             unique_id: uniqueId,
-            ...(entityId && { default_entity_id: `${HA_COMPONENT_CLIMATE}.${entityId}`, object_id: entityId }),
+            ...(entityId && entityIdFields(HA_COMPONENT_CLIMATE, entityId)),
 
             // Current temperature: reported by C-Gate as a status level on this group.
             // Template converts 0-255 C-Bus level to 0-50°C (0.5°C resolution):
@@ -590,7 +591,7 @@ class HaDiscovery {
         const payload = {
             name: null,
             unique_id: uniqueId,
-            ...(entityId && { default_entity_id: `${config.component}.${entityId}`, object_id: entityId }),
+            ...(entityId && entityIdFields(config.component, entityId)),
             state_topic: stateTopic,
             ...(!config.omitCommandTopic && { command_topic: `${MQTT_TOPIC_PREFIX_WRITE}/${networkId}/${appId}/${groupId}/${MQTT_CMD_TYPE_SWITCH}` }),
             ...config.payloads,
@@ -653,7 +654,7 @@ class HaDiscovery {
         const payload = {
             name: null,
             unique_id: uniqueId,
-            ...(entityId && { default_entity_id: `${HA_COMPONENT_BUTTON}.${entityId}_btn`, object_id: `${entityId}_btn` }),
+            ...(entityId && entityIdFields(HA_COMPONENT_BUTTON, `${entityId}_btn`)),
             command_topic: `${MQTT_TOPIC_PREFIX_WRITE}/${networkId}/${appId}/${groupId}/${MQTT_CMD_TYPE_TRIGGER}`,
             payload_press: MQTT_STATE_ON,
             qos: 0,
@@ -686,7 +687,7 @@ class HaDiscovery {
         const payload = {
             name: null,
             unique_id: uniqueId,
-            ...(entityId && { default_entity_id: `${HA_COMPONENT_SCENE}.${entityId}_scene`, object_id: `${entityId}_scene` }),
+            ...(entityId && entityIdFields(HA_COMPONENT_SCENE, `${entityId}_scene`)),
             command_topic: `${MQTT_TOPIC_PREFIX_WRITE}/${networkId}/${appId}/${groupId}/${MQTT_CMD_TYPE_SWITCH}`,
             payload_on: MQTT_STATE_ON,
             qos: 0,